fix(cli): coordinate forced process exit
